% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence.R
\name{print.sci_ci}
\alias{print.sci_ci}
\title{Print an SCI Confidence Interval}
\usage{
\method{print}{sci_ci}(x, ...)
}
\arguments{
\item{x}{An \code{sci_ci} object.}

\item{...}{Ignored.}
}
\description{
Print an SCI Confidence Interval
}
