% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{print.sci_gap}
\alias{print.sci_gap}
\title{Print an SCI Gap}
\usage{
\method{print}{sci_gap}(x, ...)
}
\arguments{
\item{x}{An \code{sci_gap} object.}

\item{...}{Ignored.}
}
\description{
Print an SCI Gap
}
