% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci_report.R
\name{print.sci_report}
\alias{print.sci_report}
\title{Print an SCI Report}
\usage{
\method{print}{sci_report}(x, ...)
}
\arguments{
\item{x}{An \code{sci_report} object.}

\item{...}{Ignored.}
}
\description{
Print an SCI Report
}
