% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sci.R
\name{print.sci_result}
\alias{print.sci_result}
\title{Print SCI Result}
\usage{
\method{print}{sci_result}(x, ...)
}
\arguments{
\item{x}{An \code{sci_result} object.}

\item{...}{Ignored.}
}
\description{
Print SCI Result
}
