% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.R
\name{print.sci_targets}
\alias{print.sci_targets}
\title{Print an SCI Target}
\usage{
\method{print}{sci_targets}(x, ...)
}
\arguments{
\item{x}{An \code{sci_targets} object.}

\item{...}{Ignored.}
}
\description{
Print an SCI Target
}
